Restructure @/components/data
