+++ apps/astro-app/src/pages/data/[databaseName]/[tableName]/graphs.astro
@@ -9,7 +9,7 @@
 import type { DatabaseInfo, TableInfo } from "src/env";
 import config from "@universal_root/config.yml";
 import Layout from "@/layouts/Layout.astro";
-import DatasetControlPanel from "@root/src/components/data/dataset-control-panel";
+import DatasetControlPanel from "@root/src/components/data/chart/dataset-control-panel";
 import LoginDialog from "@/components/auth/login-dialog";
 
 export const prerender = true;
