+++ apps/astro-app/src/pages/personal/Data-[datasetName].astro
@@ -22,7 +22,7 @@ if (!entry) {
 
 import Layout from "@/layouts/Layout.astro";
 
-import { RadialCharts } from "@root/src/components/data/radial-charts";
+import { RadialCharts } from "@root/src/components/data/chart/radial-charts";
 ---
 
 <Layout
